requirements.txt
langchain
langchain-google-genai
streamlit
openai
pydantic
python-dotenv
google-generativeai
